requirements.txt
antiorm==1.2.1
asgiref==3.7.2
boto3==1.33.6
django-storages==1.14.2
django-apscheduler==0.6.2
db==0.1.1
db-sqlite3==0.0.1
Django==4.2.4
django-taggit==5.0.1
openai==0.28.0
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3